requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==3.2.0
requests==2.31.0
google-generativeai==0.3.2 